refactor: navbars
